namig convention changed

+++ Engine/Graphics.cpp
@@ -377,8 +377,7 @@ void Graphics::DrawPoliLine(std::vector<JC_Point2d> point_data, Color Color_in)
 {
 	JC_Point2d Current;
 	JC_Point2d Previous;
-
-	
+		
 	if (point_data.size() > 1)
 	{
 		for (int i = 1; i < point_data.size(); i++)
@@ -576,19 +575,19 @@ void Graphics::DrawFlatTopTriangle(const JC_Point2d& v0, const JC_Point2d& v1, c
 	double m1 = (v2.x - v1.x) / (v2.y - v1.y);
 
 	// calculate start and end scanlines
-	const int yStart = (int)ceilf(v0.y - 0.5f);
-	const int yEnd = (int)ceilf(v2.y - 0.5f); // the scanline AFTER the last line drawn
+	const int yStart = (int)ceil(v0.y - 0.5);
+	const int yEnd = (int)ceil(v2.y - 0.5); // the scanline AFTER the last line drawn
 
 	for (int y = yStart; y < yEnd; y++)
 	{
 		// caluclate start and end points (x-coords)
 		// add 0.5 to y value because we're calculating based on pixel CENTERS
-		const float px0 = m0 * (float(y) + 0.5f - v0.y) + v0.x;
-		const float px1 = m1 * (float(y) + 0.5f - v1.y) + v1.x;
+		const double px0 = m0 * (double(y) + 0.5 - v0.y) + v0.x;
+		const double px1 = m1 * (double(y) + 0.5 - v1.y) + v1.x;
 
 		// calculate start and end pixels
-		const int xStart = (int)ceilf(px0 - 0.5f);
-		const int xEnd = (int)ceilf(px1 - 0.5f); // the pixel AFTER the last pixel drawn
+		const int xStart = (int)ceil(px0 - 0.5);
+		const int xEnd = (int)ceil(px1 - 0.5); // the pixel AFTER the last pixel drawn
 
 		for (int x = xStart; x < xEnd; x++)
 		{
@@ -604,19 +603,19 @@ void Graphics::DrawFlatBottomTriangle(const JC_Point2d& v0, const JC_Point2d& v1
 	double m1 = (v2.x - v0.x) / (v2.y - v0.y);
 
 	// calculate start and end scanlines
-	const int yStart = (int)ceilf(v0.y - 0.5f);
-	const int yEnd = (int)ceilf(v2.y - 0.5f); // the scanline AFTER the last line drawn
+	const int yStart = (int)ceil(v0.y - 0.5);
+	const int yEnd = (int)ceil(v2.y - 0.5); // the scanline AFTER the last line drawn
 
 	for (int y = yStart; y < yEnd; y++)
 	{
 		// caluclate start and end points
 		// add 0.5 to y value because we're calculating based on pixel CENTERS
-		const float px0 = m0 * (float(y) + 0.5f - v0.y) + v0.x;
-		const float px1 = m1 * (float(y) + 0.5f - v0.y) + v0.x;
+		const double px0 = m0 * (double(y) + 0.5 - v0.y) + v0.x;
+		const double px1 = m1 * (double(y) + 0.5 - v0.y) + v0.x;
 
 		// calculate start and end pixels
-		const int xStart = (int)ceilf(px0 - 0.5f);
-		const int xEnd = (int)ceilf(px1 - 0.5f); // the pixel AFTER the last pixel drawn
+		const int xStart = (int)ceil(px0 - 0.5);
+		const int xEnd = (int)ceil(px1 - 0.5); // the pixel AFTER the last pixel drawn
 
 		for (int x = xStart; x < xEnd; x++)
 		{
